move init_dirs
